+++ tests/test_plugin.py
@@ -31,7 +31,10 @@ def test_httpx_mock_unused_response(testdir):
     result = testdir.runpytest()
     result.assert_outcomes(errors=1, passed=1)
     result.stdout.fnmatch_lines(
-        ["*AssertionError: The following responses are mocked but not requested: *"]
+        [
+            "*AssertionError: The following responses are mocked but not requested:",
+            "*Match all requests",
+        ]
     )
 
 
@@ -72,7 +75,10 @@ def test_httpx_mock_unused_callback(testdir):
     result = testdir.runpytest()
     result.assert_outcomes(errors=1, passed=1)
     result.stdout.fnmatch_lines(
-        ["*AssertionError: The following callbacks are registered but not executed: *"]
+        [
+            "*AssertionError: The following responses are mocked but not requested:",
+            "*Match all requests",
+        ]
     )
 
 
@@ -94,6 +100,73 @@ def test_httpx_mock_unused_callback_without_assertion(testdir):
         
             httpx_mock.add_callback(unused)
 
+    """
+    )
+    result = testdir.runpytest()
+    result.assert_outcomes(passed=1)
+
+
+def test_httpx_mock_non_mocked_hosts_sync(testdir):
+    """
+    Non mocked hosts should go through while other requests should be mocked.
+    """
+    testdir.makepyfile(
+        """
+        import httpx
+        import pytest
+        
+        @pytest.fixture
+        def non_mocked_hosts() -> list:
+            return ["localhost"]
+
+        def test_httpx_mock_non_mocked_hosts_sync(httpx_mock):
+            httpx_mock.add_response()
+            
+            with httpx.Client() as client:
+                # Mocked request
+                client.get("http://foo.tld")
+            
+                # Non mocked request
+                with pytest.raises(httpx.ConnectError):
+                    client.get("http://localhost:5005")
+            
+            # Assert that a single request was mocked
+            assert len(httpx_mock.get_requests()) == 1
+            
+    """
+    )
+    result = testdir.runpytest()
+    result.assert_outcomes(passed=1)
+
+
+def test_httpx_mock_non_mocked_hosts_async(testdir):
+    """
+    Non mocked hosts should go through while other requests should be mocked.
+    """
+    testdir.makepyfile(
+        """
+        import httpx
+        import pytest
+        
+        @pytest.fixture
+        def non_mocked_hosts() -> list:
+            return ["localhost"]
+
+        @pytest.mark.asyncio
+        async def test_httpx_mock_non_mocked_hosts_async(httpx_mock):
+            httpx_mock.add_response()
+            
+            async with httpx.AsyncClient() as client:
+                # Mocked request
+                await client.get("http://foo.tld")
+            
+                # Non mocked request
+                with pytest.raises(httpx.ConnectError):
+                    await client.get("http://localhost:5005")
+            
+            # Assert that a single request was mocked
+            assert len(httpx_mock.get_requests()) == 1
+            
     """
     )
     result = testdir.runpytest()